requirements.txt
# Google Cloud / Vertex AI
google-adk
google-cloud-aiplatform[rag]>=1.71.0
google-cloud-storage>=2.18.0

# Gemini API (direct — no Vertex AI needed)
google-generativeai>=0.8.0

# Neo4j + Graphiti
neo4j>=5.28.0
graphiti-core>=0.12.0

# PDF processing
pypdf>=4.0.0
pymupdf>=1.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
tenacity>=8.2.0
tqdm>=4.66.0
loguru>=0.7.0
